pyproject.toml
[tool.poetry]
name = "kivydata"
version = "0.1.0"
description = ""
authors = ["AccelQuasarDragon <138998466+AccelQuasarDragon@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
kivy = "^2.3.0"
kivy-reloader = "^0.3.2"
kivy-matplotlib-widget = "^0.12.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
